requirements.txt
Flask
python-dotenv
beautifulsoup4